requirements.txt
streamlit==1.41.1
pandas==2.2.3
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
python-dotenv==1.0.1
plotly==5.24.1
